requirements.txt
mysql-connector-python~=9.3.0
